requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
pydantic-core==2.14.1
slowapi==0.1.9
redis==5.0.1
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
